subscription-api/requirements.txt
 djangorestframework
 markdown       # Markdown support for the browsable API.
 django-filter  
django-cors-headers
psycopg2-binary